# util/deephash: simplify hex encode