@crawshaw 2021-07-03
I'm not convinced saving the extra alloc is worth the upstream commit. I wanted to put in code my case that reducing aliasing improves readability.