@bradfitz 2021-07-05
I've sent https://github.com/tailscale/tailscale/pull/2327 instead which, in the process of redoing the API for use for the control server, makes this PR needed.